test: actualizar tests para Flow01, MessageStatus y smoke tests
- Flow01OrchestratorTests: agregar IVariantTrackingService al ctor
- MessageStatusServiceTests: agregar IIdempotencyService, test de idempotencia
- TC02: corregir assertWithReason en BeOneOf
- VariantStatsEndpointsTests: corregir conversion short/int en WeightPct
- SmokeTestBase: agregar helpers para assertions

+++ apps/api/tests/ClinicBoost.Tests/Features/Variants/VariantStatsEndpointsTests.cs
@@ -179,7 +179,7 @@ public class VariantStatsEndpointsTests : IAsyncLifetime
             FlowId     = "flow_01",
             TemplateId = "HXabc",
             VariantKey = "control",
-            WeightPct  = pct,
+            WeightPct  = (short)pct,
         });
 
         Assert.Equal(400, GetStatusCode(result));
